--- analysis/2_khipu_centered_ion_stats/elution_parameters_45studies_orbi.tsv
@@ -1,46 +1,0 @@
-ID	number_good_khipus	mean_m0	stdev_m0	mean_shift	stdev_shift
-ST001237_HILICpos_B2_ppm5_3524314	2895	6.473	3.651	-0.035	0.612
-ST002937_HILICpos_ST002937_pos_hilic_batch_1_zip_B2_ppm5_351419	4772	8.336	10.07	-0.016	0.787
-ST002112_HILICpos_B2_ppm5_3545123	3160	8.208	7.124	-0.029	0.788
-MTBLS4187_HILICpos_HILIC_ppm5_3521045	2236	6.866	10.607	-0.035	0.605
-MTBLS3852_HILICpos__ppm5_3533848	315	4.975	19.342	-0.008	0.678
-MTBLS1465_HILICpos__ppm5_3505731	2192	10.135	16.215	-0.01	0.658
-ST001335_HILICpos__ppm5_35132731	1560	6.238	3.426	-0.063	0.63
-ST001736_RPpos__ppm5_34232333	941	8.501	5.845	-0.019	0.716
-MTBLS136_RPpos_B17_ppm5_3583756	1351	4.421	1.165	-0.01	0.522
-ST002200_RPpos_17min_B3_ppm5_3422144	3365	18.187	6.085	-0.108	0.903
-ST001181_RPpos_B11_ppm5_36163417	261	9.983	18.02	-0.067	0.896
-ST002443_RPpos_B3_ppm5_81311338	396	28.957	20.819	-0.078	1.035
-ST002118_RPpos_B1_ppm5_36115624	1253	31.055	20.296	-0.158	0.994
-ST003315_RPpos__ppm5_813163714	910	5.88	4.954	-0.004	0.481
-ST002112_RPpos_B3_ppm5_356194	5977	7.451	2.792	-0.03	0.71
-MTBLS4187_RPpos__ppm5_3522135	6058	6.292	1.414	-0.023	0.623
-ST002049_RPpos_B1_ppm5_352018	1729	30.636	9.487	-0.123	1.02
-MTBLS2245_RPpos_Method 2_B1_ppm5_81318250	1095	4.708	0.903	-0.006	0.516
-ST001430_RPpos_B1_ppm5_3423140	3348	7.861	10.38	-0.038	0.579
-MTBLS205_RPpos_B1_ppm5_3531210	1039	4.485	1.012	-0.031	0.518
-ST001243_RPpos__ppm5_35185010	1070	8.757	2.716	-0.036	0.714
-ST002521_RPpos_B3_ppm5_3662420	7220	8.895	2.167	-0.028	0.685
-MTBLS2245_HILICneg_Method 4_B5_ppm5_81318755	897	4.98	0.734	-0.022	0.518
-ST002937_HILICneg_ST002937_neg_hilic_batch_1_zip_B3_ppm5_353747	3659	8.186	9.818	-0.026	0.81
-ST002112_HILICneg_B3_ppm5_3552830	2825	7.787	7.23	0.019	0.788
-MTBLS1465_HILICneg__ppm5_35195	675	9.761	6.085	-0.031	0.728
-MTBLS205_HILICneg__ppm5_353536	720	4.442	2.346	-0.033	0.55
-ST002331_RPneg_B5_ppm5_3655219	697	7.299	7.274	-0.022	0.657
-MTBLS476_RPneg__ppm5_3551151	6772	7.791	2.102	-0.024	0.695
-MTBLS4294_RPneg__ppm5_3542818	2811	10.653	5.973	-0.04	0.777
-ST002200_RPneg_5min_B4_ppm5_3423035	804	10.233	3.209	-0.049	0.679
-ST001430_RPneg_B4_ppm5_350239	2349	7.763	4.167	-0.07	0.601
-ST002238_RPneg__ppm5_36104557	2602	10.751	4.03	-0.021	0.822
-ST002521_RPneg_B2_ppm5_3673943	4045	9.871	6.516	-0.039	0.676
-ST002112_RPneg_B2_ppm5_3572946	5882	7.725	4.678	-0.031	0.715
-ST001736_RPneg__ppm5_34232652	917	10.401	8.837	-0.002	0.687
-ST002554_RPneg__ppm5_3611102	2431	10.237	9.35	0.014	0.847
-MTBLS1465_RPneg__ppm5_3514212	1742	2.378	4.885	-0.034	0.537
-MTBLS4187_RPneg__ppm5_3524019	1972	6.295	2.424	-0.014	0.566
-MTBLS205_RPneg__ppm5_3531433	1760	3.918	1.313	-0.002	0.53
-ST001240_RPneg__ppm5_35193041	1277	7.916	3.266	0.035	0.735
-ST003315_RPneg__ppm5_813164016	1152	5.992	4.795	-0.014	0.577
-ST000919_RPneg_B1_ppm5_363371	6850	5.492	1.546	-0.017	0.618
-ST001243_RPneg__ppm5_3518535	532	9.715	3.0	-0.003	0.742
-ST001335_RPneg__ppm5_35133334	1071	5.933	1.247	0.01	0.52
