--- analysis/2_khipu_centered_ion_stats/top_frequent_delta_mz_orbi_neg-20241211.tsv
@@ -1,127 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0033	4237	1.0033	M(C13)	{'C': -1, '(C13)': 1}
-67.9874	823	67.9874	NaCOOH	{'C': 1, 'O': 2, 'Na': 1, 'H': 1}
-14.0156	695	14.0156	± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)	{'C': 1, 'H': 2}
-2.0155	568	2.0155	± 2H, opening or forming of double bond	{'H': 2}
-82.0029	432
-15.9948	414	15.9948	± O, e.g. oxidation/reduction	{'O': 1}
-1.9969	410	1.9969	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-43.9898	394	43.9898	± CO2	{'C': 1, 'O': 2}
-18.0105	372	18.0105	H2O	{'H': -2, 'O': -1}
-11.9999	358	11.9999	M+H+Na[2+]	{'H': 1, 'Na': 1}
-30.0105	352
-26.0156	347
-46.0054	340	46.0054	H2O+CO	{'C': -1, 'H': -2, 'O': -2}
-28.0312	328	28.0312	± C2H4, natural alkane chains such as fatty acids	{'C': 2, 'H': 4}
-44.0261	305
-27.9949	297	27.9949	± CO	{'C': 1, 'O': 1}
-21.9819	296	21.9819	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-10.0287	281
-1.9957	271	1.9957	F <-> OH, halogen exchange with hydroxy group (typically -F + OH)	{'F': 1, 'O': -1, 'H': -1}
-66.984	269	66.984	M-HCOONa+H[1+]	{'C': -1, 'O': -2, 'Na': -1}
-23.9999	264	23.9999	H+Na	{'Na': 1, 'H': 1}
-13.9792	255	13.9792	O <-> 2H, e.g. Oxidation follwed by H2O elimination	{'H': -2, 'O': 1}
-42.0105	250	42.0105	± COCH2	{'C': 2, 'O': 1, 'H': 2}
-16.0312	238
-2.006	234	2.006	M(C13)+H[1+]	{'C': -1, '(C13)': 1, 'H': 1}
-60.021	229
-135.9748	224	135.9748	2X NaCOOH	{'C': 2, 'O': 4, 'H': 2, 'Na': 2}
-2.0042	223
-15.9738	217	15.9738	Na+<-> K+, salt adduct	{'Na': -1, 'K': 1}
-1.9792	215	1.9792	K+ <-> Cl-+2H2+, salt adduct	{'K': -1, 'Cl': 1, 'H': 2}
-1.9977	211	1.9977	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-3.9948	208
-32.0261	195	32.0261	MeOH	{'C': 1, 'H': 4, 'O': 1}
-9.9842	194
-4.0312	193
-57.9587	190	57.9587	NaCl	{'Na': 1, 'Cl': 1}
-15.0189	182
-46.0418	178
-58.0418	177	58.0418	"+C3H6O, acetone condensation"	{'C': 3, 'H': 6, 'O': 1}
-21.9842	175	21.9842	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-3.0189	172	3.0189	3H	{'H': 3}
-13.0122	168
-42.0469	167	42.0469	± C3H6, propylation	{'C': 3, 'H': 6}
-56.0261	165
-68.9908	165	68.9908	M+HCOONa[1+]	{'C': 1, 'O': 2, 'Na': 2, 'H': 1}
-31.9898	162	31.9898	± 2O, oxygen loss	{'O': 2}
-58.0054	160	58.0054	± CO2CH2	{'C': 2, 'O': 2, 'H': 2}
-20.0261	160
-30.0469	160
-27.0189	157
-1.0121	155	1.0121	-H[-]	{'H': -1}
-39.9949	154	39.9949	Na+ <-> -H2O+H	{'Na': 1, 'O': 1, 'H': 1}
-70.0418	154
-12.0363	153
-47.0088	153
-25.9792	153
-6.0169	152
-164.0058	152
-38.0156	149
-10.0206	148
-40.0312	146	40.0312	"(C3H6O - H2O), acetone condensation after dehydration"	{'H': 4, 'C': 3}
-72.0575	144
-80.9996	143
-54.0105	143
-8.005	143
-6.0105	141
-74.0367	137
-62.0003	133
-72.0211	133
-50.0156	132
-8.0317	131
-34.0054	130
-29.9741	126	29.9741	NO2 <-> NH2, nitro compounds	{'H': -2, 'O': 2}
-18.0468	126
-56.0625	125	56.0625	± C4H8, butylation	{'C': 4, 'H': 8}
-11.9635	125
-35.9766	125	35.9766	Cl-+2H2+ <-> H+, salt adduct	{'Cl': 1, 'H': 1}
-52.0312	124
-16.9901	124
-0.5015	123	0.5015	double charged with C13	{'C': -1, '(C13)': 1}
-88.0523	123
-0.9935	121	0.9935	M(Cl37)-H[-]	{'Cl': -1, '(Cl37)': 1, 'H': -1}
-2.9996	121	2.9996	M(S34)+H[1+]	{'S': -1, '(S34)': 1, 'H': 1}
-48.021	121
-35.9999	120
-54.0468	120
-29.9895	119
-62.0367	119
-68.0261	117
-78.0164	117
-60.0574	117
-0.9951	117
-97.9769	117	97.9769	± H3PO4, phosphorous compounds	{'O': 4, 'H': 3, 'P': 1}
-34.0418	116
-1.9945	115	1.9945	F <-> OH, halogen exchange with hydroxy group (typically -F + OH)	{'F': 1, 'O': -1, 'H': -1}
-24.0443	114
-44.9931	114
-5.9894	114
-7.9686	113
-52.0134	112
-67.9897	112	67.9897	NaCOOH	{'C': 1, 'O': 2, 'Na': 1, 'H': 1}
-61.9704	111
-90.0316	111
-77.9772	110
-86.0366	110
-41.0265	110	41.0265	Acetonitrile	{'C': 2, 'H': 3, 'N': 1}
-82.0053	108
-22.0418	107
-70.0054	107
-83.0064	106
-0.9994	105	0.9994	M(C13)	{'C': -1, '(C13)': 1}
-17.9741	105
-27.0109	104	27.0109	± HCN, nitrile compounds	{'N': 1, 'H': 1, 'C': 1}
-16.9982	103
-76.0159	103
-83.9612	103	83.9612	KCOOH	{'O': 2, 'C': 1, 'K': 1, 'H': 1}
-14.9914	103	14.9914	M-H+O[-]	{'H': -1, 'O': 1}
-6.0469	102
-96.0186	102
-19.9897	101
-22.0206	101
-0.984	101	0.984	OH <-> NH2, e.g. de-amidiation, CHNO compounds	{'O': 1, 'H': -1, 'N': -1}
-19.0139	100	19.0139	M-H2O-H[-]	{'H': -3, 'O': -1}
-84.0575	100
-27.0278	100
-48.0575	100
