--- analysis/2_khipu_centered_ion_stats/top_frequent_delta_mz_orbi_pos-20241211.tsv
@@ -1,171 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0033	4156	1.0033	M(C13)	{'C': -1, '(C13)': 1}
-14.0155	899	14.0155	± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)	{'C': 1, 'H': 2}
-18.0104	883	18.0104	H2O	{'H': -2, 'O': -1}
-2.0155	717	2.0155	± 2H, opening or forming of double bond	{'H': 2}
-21.9818	656	21.9818	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-44.0261	652
-28.0312	621	28.0312	± C2H4, natural alkane chains such as fatty acids	{'C': 2, 'H': 4}
-15.9948	479	15.9948	± O, e.g. oxidation/reduction	{'O': 1}
-17.0264	451	17.0264	NH3	{'N': -1, 'H': -3}
-0.5015	449	0.5015	double charged with C13	{'C': -1, '(C13)': 1}
-26.0155	439
-27.9947	433	27.9947	± CO	{'C': 1, 'O': 1}
-11.9999	425	11.9999	M+H+Na[2+]	{'H': 1, 'Na': 1}
-30.0104	369
-32.026	350	32.026	MeOH	{'C': 1, 'H': 4, 'O': 1}
-42.0104	340	42.0104	± COCH2	{'C': 2, 'O': 1, 'H': 2}
-67.9872	325	67.9872	NaCOOH	{'C': 1, 'O': 2, 'Na': 1, 'H': 1}
-13.9791	319	13.9791	O <-> 2H, e.g. Oxidation follwed by H2O elimination	{'H': -2, 'O': 1}
-23.9998	317	23.9998	H+Na	{'Na': 1, 'H': 1}
-46.0053	313	46.0053	H2O+CO	{'C': -1, 'H': -2, 'O': -2}
-42.0468	313	42.0468	± C3H6, propylation	{'C': 3, 'H': 6}
-16.0312	313
-88.0522	304
-41.0263	294	41.0263	Acetonitrile	{'C': 2, 'H': 3, 'N': 1}
-15.9738	290	15.9738	Na+<-> K+, salt adduct	{'Na': -1, 'K': 1}
-22.9852	281	22.9852	Na	{'Na': 1}
-30.0468	266
-4.0312	258
-56.0624	257	56.0624	± C4H8, butylation	{'C': 4, 'H': 8}
-60.0209	256
-9.9842	255
-1.9792	244	1.9792	K+ <-> Cl-+2H2+, salt adduct	{'K': -1, 'Cl': 1, 'H': 2}
-3.9948	244
-58.0417	240	58.0417	"+C3H6O, acetone condensation"	{'C': 3, 'H': 6, 'O': 1}
-4.9553	237	4.9553	Na+<-> NH4+, salt adduct	{'Na': 1, 'N': -1, 'H': -4}
-19.0138	227	19.0138	M-H2O-H[-]	{'H': -3, 'O': -1}
-36.0209	226
-46.0417	224
-15.0189	220
-40.0311	216	40.0311	"(C3H6O - H2O), acetone condensation after dehydration"	{'H': 4, 'C': 3}
-2.0065	214	2.0065	M(C13)+H[1+]	{'C': -1, '(C13)': 1, 'H': 1}
-17.007	209	17.007	M-H2O+H[1+]	{'H': -1, 'O': -1}
-13.0122	209
-1.9968	206	1.9968	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-20.0261	206
-3.0189	203	3.0189	3H	{'H': 3}
-45.0577	196
-58.0052	192	58.0052	± CO2CH2	{'C': 2, 'O': 2, 'H': 2}
-72.0573	191
-21.9841	188	21.9841	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-56.026	188
-31.9897	185	31.9897	± 2O, oxygen loss	{'O': 2}
-10.0207	183
-54.0467	180
-12.0363	180
-43.9896	180	43.9896	± CO2	{'C': 1, 'O': 2}
-43.0227	178
-60.0573	172
-74.0365	171
-72.0209	171
-22.0131	169
-15.0107	169
-1.0122	169	1.0122	-H[-]	{'H': -1}
-2.0042	164
-70.0781	164
-39.9924	162	39.9924	Na+ <-> -H2O+H	{'Na': 1, 'O': 1, 'H': 1}
-8.005	160
-39.9947	159	39.9947	Na+ <-> -H2O+H	{'Na': 1, 'O': 1, 'H': 1}
-70.0417	159
-38.0154	159
-1.001	158	1.001	M(C13)	{'C': -1, '(C13)': 1}
-34.0053	158
-45.0294	158
-35.037	157	35.037	NH3+H2O	{'H': -5, 'O': -1, 'N': -1}
-74.073	156
-25.9791	156
-0.9839	155	0.9839	OH <-> NH2, e.g. de-amidiation, CHNO compounds	{'O': 1, 'H': -1, 'N': -1}
-27.0107	155	27.0107	± HCN, nitrile compounds	{'N': 1, 'H': 1, 'C': 1}
-48.0209	153
-18.0468	153
-0.997	153
-1.9979	151	1.9979	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-27.0278	149
-29.0345	149
-6.0105	148
-52.0311	146
-44.0624	145
-132.0785	144
-0.9951	143
-29.974	140	29.974	NO2 <-> NH2, nitro compounds	{'H': -2, 'O': 2}
-14.0519	140
-12.9951	139
-62.0365	139
-37.9557	139	37.9557	K-H	{'K': 1, 'H': -1}
-54.0104	139
-84.0937	137
-34.0417	132
-102.0679	132
-82.0028	131
-11.9635	130
-29.0264	128
-57.9584	128	57.9584	NaCl	{'Na': 1, 'Cl': 1}
-3.0108	126
-23.0158	126
-16.023	125	16.023	M-NH3+H[1+]	{'N': -1, 'H': -2}
-59.0369	125	59.0369	± C2H5NO, N-acetyl or cut from side chain of Asn in peptides	{'C': 2, 'H': 5, 'N': 1, 'O': 1}
-50.0154	124
-26.0519	124
-7.9685	124
-17.9741	124
-35.9998	124
-86.0365	124
-50.0366	123
-6.0468	123
-42.0297	122
-43.042	122
-24.0363	122
-1.9956	121	1.9956	F <-> OH, halogen exchange with hydroxy group (typically -F + OH)	{'F': 1, 'O': -1, 'H': -1}
-32.0624	121
-58.078	121
-86.073	120
-18.0298	120	18.0298	NH4+	{'N': 1, 'H': 4}
-5.9893	118
-1.0077	118	1.0077	-H[-]	{'H': -1}
-64.0522	117
-22.0418	117
-68.0624	117
-27.0188	117
-13.0315	116
-14.9914	116	14.9914	M-H+O[-]	{'H': -1, 'O': 1}
-48.0573	116
-8.0261	116
-2.9995	115	2.9995	M(S34)+H[1+]	{'S': -1, '(S34)': 1, 'H': 1}
-61.0525	115
-10.0287	114
-45.0212	114
-31.042	114
-55.9896	112
-68.026	112
-25.0122	111
-76.0522	111
-22.0206	111
-19.9685	110
-19.9897	109
-39.0107	109
-82.078	107
-26.9995	107
-135.9746	107	135.9746	2X NaCOOH	{'C': 2, 'O': 4, 'H': 2, 'Na': 2}
-38.0519	106
-43.9638	106	43.9638	2Na-2H	{'Na': 2, 'H': -2}
-70.0052	106
-66.0467	106
-3.9584	105
-1.0203	104
-22.0053	104
-84.0572	104
-24.9951	104	24.9951	CN <-> H, nitrile compounds	{'C': 1, 'N': 1, 'H': -1}
-31.0056	103
-11.0158	102
-90.0678	102
-14.9996	102
-64.0158	102
-1.0315	102
-1.0062	101	1.0062	-H[-]	{'H': -1}
-10.9965	101
-5.974	101
-19.042	100
-7.9897	100
-5.9587	100
-76.0159	100
