--- analysis/2_khipu_centered_ion_stats/top_frequent_delta_mz_tof_neg-20241211.tsv
@@ -1,27 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0033	244	1.0033	M(C13)	{'C': -1, '(C13)': 1}
-67.9873	47	67.9873	NaCOOH	{'C': 1, 'O': 2, 'Na': 1, 'H': 1}
-2.0154	26	2.0154	± 2H, opening or forming of double bond	{'H': 2}
-82.0027	19
-135.9746	18	135.9746	2X NaCOOH	{'C': 2, 'O': 4, 'H': 2, 'Na': 2}
-43.9898	17	43.9898	± CO2	{'C': 1, 'O': 2}
-1.007	17	1.007	-H[-]	{'H': -1}
-66.9838	14	66.9838	M-HCOONa+H[1+]	{'C': -1, 'O': -2, 'Na': -1}
-2.0059	14	2.0059	M(C13)+H[1+]	{'C': -1, '(C13)': 1, 'H': 1}
-14.0155	13	14.0155	± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)	{'C': 1, 'H': 2}
-61.9702	13
-68.9905	12	68.9905	M+HCOONa[1+]	{'C': 1, 'O': 2, 'Na': 2, 'H': 1}
-0.9991	12	0.9991	M(C13)	{'C': -1, '(C13)': 1}
-26.0154	12
-0.9963	11
-74.0363	11
-1.997	11	1.997	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-46.0052	11	46.0052	H2O+CO	{'C': -1, 'H': -2, 'O': -2}
-129.9575	11
-1.0105	10	1.0105	-H[-]	{'H': -1}
-24.0	10	24.0	H+Na	{'Na': 1, 'H': 1}
-60.0212	10
-1.9979	10	1.9979	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-10.0289	10
-21.9819	10	21.9819	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-18.0107	10	18.0107	± H2O, water addition/loss	{'H': 2, 'O': 1}
