--- analysis/2_khipu_centered_ion_stats/top_frequent_delta_mz_tof_pos-20241211.tsv
@@ -1,55 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0032	232	1.0032	M(C13)	{'C': -1, '(C13)': 1}
-0.9992	42	0.9992	M(C13)	{'C': -1, '(C13)': 1}
-1.0062	38	1.0062	-H[-]	{'H': -1}
-18.0106	38	18.0106	± H2O, water addition/loss	{'H': 2, 'O': 1}
-21.9821	32	21.9821	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-14.0155	31	14.0155	± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)	{'C': 1, 'H': 2}
-1.0073	30	1.0073	-H[-]	{'H': -1}
-0.9972	28
-2.0154	26	2.0154	± 2H, opening or forming of double bond	{'H': 2}
-0.5015	26	0.5015	double charged with C13	{'C': -1, '(C13)': 1}
-17.0265	20	17.0265	NH3	{'N': -1, 'H': -3}
-28.0311	20	28.0311	± C2H4, natural alkane chains such as fatty acids	{'C': 2, 'H': 4}
-1.0096	20	1.0096	-H[-]	{'H': -1}
-1.9972	19	1.9972	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-22.985	18	22.985	Na	{'Na': 1}
-1.0123	18
-1.9982	18	1.9982	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-1.0116	18	1.0116	-H[-]	{'H': -1}
-12.0	17	12.0	M+H+Na[2+]	{'H': 1, 'Na': 1}
-2.0063	17	2.0063	M(C13)+H[1+]	{'C': -1, '(C13)': 1, 'H': 1}
-46.0055	16	46.0055	± CO+H2O (carboxylic acid)	{'C': 1, 'O': 2, 'H': 2}
-74.0184	16
-21.9837	16	21.9837	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-44.0262	15
-16.0313	15
-26.0155	15
-26.0157	15
-2.0051	14
-21.9848	14	21.9848	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-21.9805	13	21.9805	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-27.9948	13	27.9948	± CO	{'C': 1, 'O': 1}
-15.995	13	15.995	± O, e.g. oxidation/reduction	{'O': 1}
-2.0142	12	2.0142	2H	{'H': 2}
-23.9998	12	23.9998	H+Na	{'Na': 1, 'H': 1}
-32.026	12	32.026	MeOH	{'C': 1, 'H': 4, 'O': 1}
-2.0039	11
-39.9926	11	39.9926	Na+ <-> -H2O+H	{'Na': 1, 'O': 1, 'H': 1}
-42.0466	11	42.0466	± C3H6, propylation	{'C': 3, 'H': 6}
-4.9554	11	4.9554	Na+<-> NH4+, salt adduct	{'Na': 1, 'N': -1, 'H': -4}
-18.0121	11	18.0121	± H2O, water addition/loss	{'H': 2, 'O': 1}
-2.0013	11	2.0013	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-2.017	11	2.017	± 2H, opening or forming of double bond	{'H': 2}
-2.0006	11	2.0006	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
-3.0184	11	3.0184	3H	{'H': 3}
-56.0624	10	56.0624	± C4H8, butylation	{'C': 4, 'H': 8}
-9.9842	10
-30.0106	10
-0.9924	10	0.9924	M(Cl37)-H[-]	{'Cl': -1, '(Cl37)': 1, 'H': -1}
-1.9961	10	1.9961	M(S34)	{'S': -1, '(S34)': 1}
-21.9857	10	21.9857	Na+<-> H+, salt adduct	{'Na': 1, 'H': -1}
-0.9926	10	0.9926	M(Cl37)-H[-]	{'Cl': -1, '(Cl37)': 1, 'H': -1}
-0.9928	10	0.9928	M(Cl37)-H[-]	{'Cl': -1, '(Cl37)': 1, 'H': -1}
-3.9948	10
-15.9737	10	15.9737	Na+<-> K+, salt adduct	{'Na': -1, 'K': 1}
