+++ analysis/3_get_stats_rtbins/run_stats_by_batches_tof.py
@@ -194,7 +194,7 @@ def explain_a_dataset_byMS2(all_features, ms2_tree, rt_stdev, limit_ppm=5):
             
             if matched_ms2_mzs:
                 # _deltas = [sp['mz']-x for x in matched_ms2_mzs] # delta between ms2 peaks and precursors
-                matched.append((ff['id_number'], sp['name'], matched_ms2_mzs))
+                matched.append((ff['id_number'], ff['mz'], sp['name'], matched_ms2_mzs))
                 
     return _have_precursors, matched
 
@@ -275,51 +275,87 @@ for line in open('elution_parameters_16studies_tof.tsv').readlines()[1:]:
     dict_rtwindow[a[0]] = float(a[5])
     
 
-candidate_fragments = '''14.0155	792	14.0155	['14.01565', "± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)", "{'C': 1, 'H': 2}"]
-18.0104	780	18.0104	['-18.010565', 'H2O', "{'H': -2, 'O': -1}"]
-2.0156	634	2.0156	['2.01565', '± 2H, opening or forming of double bond', "{'H': 2}"]
-28.0312	550	28.0312	['28.0313', '± C2H4, natural alkane chains such as fatty acids', "{'C': 2, 'H': 4}"]
-15.9948	420	15.9948	['15.99492', '± O, e.g. oxidation/reduction', "{'O': 1}"]
-17.0264	404	17.0264	['-17.026549', 'NH3', "{'N': -1, 'H': -3}"]
-26.0155	392	26.0155	[' C2H2']
-27.9948	385	27.9948	['27.99492', '± CO', "{'C': 1, 'O': 1}"]
-32.026	311	32.0261	['32.026215', 'MeOH', "{'C': 1, 'H': 4, 'O': 1}"]
-42.0104	301	42.0104	['42.01057', '± COCH2', "{'C': 2, 'O': 1, 'H': 2}"]
-67.9872	295	67.9873	['67.987424', 'NaCOOH', "{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"]
-13.9791	287	13.9791	['13.97927', 'O <-> 2H, e.g. Oxidation follwed by H2O elimination', "{'H': -2, 'O': 1}"]
-42.0468	278	42.0468	['42.04695', '± C3H6, propylation', "{'C': 3, 'H': 6}"]
-46.0053	277	46.0053	['-46.005479', 'H2O+CO', "{'C': -1, 'H': -2, 'O': -2}"]
+pos_candidate_fragments = '''18.01	113	18.010565	water	{'H': 2, 'O': 1}
+14.015	83	14.015649	addition of acetic acid and loss of CO2. Reaction: (+C2H2O2) and (-CO2)	{'C': 1, 'H': 2}
+2.015	80	2.014552	2H	{'H': 2}
+28.0305	59	28.0313	± C2H4, natural alkane chains such as fatty acids	{'C': 2, 'H': 4}
+22.9845	54	22.989218	 Na	  { 'Na':1 }
+46.0055	54	46.00548	± CO+H2O (carboxylic acid)	{'C': 1, 'O': 2, 'H': 2}
+17.0265	53	17.0265	addition of ammonia. Reaction: (+NH3)	{'N': 1, 'H': 3}
+11.9995	45	12.0	methylation and reduction	{'C': 1}
+44.0255	41	44.0262	hydroxyethylation	{'C': 2, 'H': 4, 'O': 1}
+26.015	41	26.01565	acetylation and loss of oxygen. Reaction: (+C2H2O) and (-O)	{'C': 2, 'H': 2}
+15.995	40	15.99492	± O, e.g. oxidation/reduction	{'O': 1}
+16.031	39	16.0313	Methylation + reduction	{'C': 1, 'H': 4}
+32.026	37	32.026215	MeOH	{'C': 1, 'H': 4, 'O': 1}
+39.993	37	39.9925	extra OH sodium adduct	{'H': 1, 'O': 1}
+27.9945	37	27.9949	addition of CO. Reaction: (+CO)	{'C': 1, 'O': 1}
+23.999	36	24.0	acetylation and loss of water. Reaction: (+C2H2O) and (-H2O)	{'C': 2}
+42.0465	35	42.04695	± C3H6, propylation	{'C': 3, 'H': 6}
+9.984	32	9.98435	addition of CO and loss of water. Reaction: (+CO) and (-H2O)	{'C': 1, 'H': -2}
+30.0105	31	30.010564	addition of acetic acid and loss of CO. Reaction: (+C2H2O2) and (-CO)	{'C': 1, 'H': 2, 'O': 1}
+56.0625	30	56.0626	± C4H8, butylation	{'C': 4, 'H': 8}
 '''
-candidate_fragments = [
-    (float(x.split()[0]), x) for x in candidate_fragments.splitlines()
+pos_candidate_fragments = [
+    (float(x.split()[0]), x) for x in pos_candidate_fragments.splitlines()
 ]
-isf_candidate_fragments = [x[0] for x in candidate_fragments]
-len(candidate_fragments), candidate_fragments[3]
+pos_isf_candidate_fragments = [x[0] for x in pos_candidate_fragments]
+
+neg_candidate_fragments = '''67.9875	140	67.987424	NaCOOH	"{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"
+2.015	70	2.014552	2H	{'H': 2}
+135.974	49	135.974848	2X NaCOOH	"{'C': 2, 'O': 4, 'H': 2, 'Na': 2}"
+82.002	48	82.003035	methylation and addition of trifluoromethyl. Reaction: (+CH2) and (+CF3-H)	"{'C': 2, 'H': 1, 'F': 3}"
+1.011	41	1.007276467	1H	{'H':1 }
+0.996	39	0.996585	addition of Guanine and loss of D-ribose. Reaction: (+C5H3N5) and (-C5H8O4)	"{'H': -5, 'N': 5, 'O': -4}"
+43.989	35	43.9898	addition of CO2. Reaction: (+CO2)	"{'C': 1, 'O': 2}"
+14.015	32	14.015649	addition of acetic acid and loss of CO2. Reaction: (+C2H2O2) and (-CO2)	"{'C': 1, 'H': 2}"
+46.005	31	46.005305	addition of Phosphate and dechlorination. Reaction: (+HPO3) and (-Cl+H)	"{'H': 2, 'O': 3, 'P': 1, 'Cl': -1}"
+26.015	29	26.01565	acetylation and loss of oxygen. Reaction: (+C2H2O) and (-O)	"{'C': 2, 'H': 2}"
+61.97	29	61.975755	addition of Phosphate and defluorination. Reaction: (+HPO3) and (-F+H)	"{'H': 2, 'O': 3, 'P': 1, 'F': -1}"
+129.957	29	129.958482	addition of di-phosphate and denitrification. Reaction: (+H2P2O6) and (NO2 -> NH2)	"{'H': 4, 'O': 4, 'P': 2}"
+60.021	29	-60.0211	desmolysis	"{'C': -2, 'H': -4, 'O': -2}"
+74.0365	28	74.03678	propionylation	"{'C': 3, 'H': 6, 'O': 2}"
+23.9995	27	24	acetylation and loss of water. Reaction: (+C2H2O) and (-H2O)	{'C': 2}
+10.029	27	10.028802	addition of CO2 and dechlorination. Reaction: (+CO2) and (-Cl+H)	"{'C': 1, 'H': 1, 'O': 2, 'Cl': -1}"
+44.026	26	44.0262	hydroxyethylation	"{'C': 2, 'H': 4, 'O': 1}"
+18.0105	26	18.010565	water	"{'H': 2, 'O': 1}"
+6.0165	24	6.010565	addition of tiglyl and loss of phenyl. Reaction: (+C5H6O) and (-C6H5+H)	"{'C': -1, 'H': 2, 'O': 1}"
+15.994	23	15.9949	oxidation	{'O': 1}
+'''
+neg_candidate_fragments = [
+    (float(x.split()[0]), x) for x in neg_candidate_fragments.splitlines()
+]
+
+neg_isf_candidate_fragments = [x[0] for x in neg_candidate_fragments]
 
 # pos
 isp_pos = [ (1.003355, '13C/12C', (0, 0.8)),
                             (2.00671, '13C/12C*2', (0, 0.8)),
-                            (3.9948, '44Ca/40Ca', (0, 0.1)), # 2%
+                            # (3.9948, '44Ca/40Ca', (0, 0.1)), # 2%
                             (1.9970, '37Cl/35Cl', (0.1, 0.8)), # 24.24%
                             ]
 
 asp_pos = [  # initial patterns are relative to M+H+
                             (21.98194, 'Na/H'),
+                            (17.0265, 'NH3'),
                             (41.026549, 'ACN'),     # Acetonitrile
                             (67.987424, 'NaCOOH'),
-                            (37.955882, 'K/H'),
+                            (32.026215, 'MeOH')
                             ]
 
 # neg 
 isp_neg = [ (1.003355, '13C/12C', (0, 0.8)),
                             (2.00671, '13C/12C*2', (0, 0.8)),
                             (1.9970, '37Cl/35Cl', (0.1, 0.8)), # 24.24%
+                            (1.9958, '32S/34S', (0, 0.1)), # 4%
                             ]
 
 asp_neg = [  # initial patterns are relative to M+H+
-                            (21.98194, 'Na/H'), (67.987424, 'NaCOOH'),
+                            (21.98194, 'Na/H'), 
+                            (67.987424, 'NaCOOH'),
+                            (135.974848, 'NaCOOH*2'),
                             (82.0030, 'C2HF3'),
-                            (1.99566, 'F <-> OH'), 
+                            # (1.99566, 'F <-> OH'), 
                             ]
 
 
@@ -356,7 +392,9 @@ def process_dataset(f):
         print(f"Dataset {f} step 2 finished")
         # Step 3: Explain by mz deltas
         explained_khipu_ids, explained_feature_ids, delta_values_used = explain_a_dataset_by_mz_deltas(
-            list_khipus, remaining_features, isf_candidate_fragments, 
+            list_khipus, 
+            remaining_features, 
+            pos_isf_candidate_fragments if ion_mode == 'pos' else neg_isf_candidate_fragments, 
             rt_stdev=dict_rtwindow[f]
         )
         print(f"Dataset {f} step 3 finished")
@@ -368,7 +406,7 @@ def process_dataset(f):
         )
         delta_values_ms2 = []
         for x in matched2:
-            delta_values_ms2 += x[2]
+            delta_values_ms2.append(x)
         print(f"Dataset {f} step 4 finished")
         # Step 5: Compile results
         results = {
@@ -384,7 +422,7 @@ def process_dataset(f):
         }
         
         # Save results to JSON file
-        with open(f'output/tof/12172024_2/{ion_mode}/{f}.json', 'w', encoding='utf-8') as json_file:
+        with open(f'output/tof/12272024/{ion_mode}/{f}.json', 'w', encoding='utf-8') as json_file:
             json.dump(results, json_file, ensure_ascii=False, indent=2) 
         
         print(f"Dataset {f} processed in {time.time() - start_time:.2f} seconds")
@@ -397,8 +435,9 @@ def process_dataset(f):
 if __name__ == "__main__":
     tof_datasets = [x.rstrip() for x in open('selected_16_tof_datasets.txt').readlines()]
     # pos_orbi_datasets = [x for x in orbi_datasets if 'pos' in x]
+    # neg_orbi_datasets = [x for x in tof_datasets if 'neg' in x]
     # Set number of workers to number of available CPUs
-    max_workers = 8  # Adjust based on available cores or system capacity
+    max_workers = 9  # Adjust based on available cores or system capacity
 
     # List to track completed results
     tally_pos = []
@@ -409,6 +448,9 @@ if __name__ == "__main__":
         
         # Process results as they complete
         for future in as_completed(futures):
+            active_processes = sum(1 for f in futures if not f.done())
+            print(f"Currently active processes: {active_processes}")
+            
             dataset_name = futures[future]
             try:
                 result = future.result()
