+++ analysis/3_get_stats_rtbins/top_frequent_delta_mz_tof_neg_20241227.tsv
@@ -1,0 +1,34 @@
+delta_mz	count_estimate	mass_signature	note	dict
+1.003	595	1.00335	C isotope	"{'(C12)': -1, '(C13)': 1}"
+67.9875	140	67.987424	NaCOOH	"{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"
+2.015	70	2.014552	2H	{'H': 2}
+135.974	49	135.974848	2X NaCOOH	"{'C': 2, 'O': 4, 'H': 2, 'Na': 2}"
+82.002	48	82.003035	methylation and addition of trifluoromethyl. Reaction: (+CH2) and (+CF3-H)	"{'C': 2, 'H': 1, 'F': 3}"
+2.006	47	1.99705	Cl isotope	"{'(Cl35)': -1, '(Cl37)': 1}"
+66.984	43	-66.98012353	M+NaCOOH-H[1+]	"{'C':1 , 'O':2,'Na':1 ,}"
+1.9985	42	1.9958	S isotope	"{'(S32)': -1, '(S34)': 1}"
+1.011	41	1.007276467	1H	{'H':1 }
+0.996	39	0.996585	addition of Guanine and loss of D-ribose. Reaction: (+C5H3N5) and (-C5H8O4)	"{'H': -5, 'N': 5, 'O': -4}"
+21.982	36	21.981943	Sodium adduct	"{'H': -1, 'Na': 1}"
+43.989	35	43.9898	addition of CO2. Reaction: (+CO2)	"{'C': 1, 'O': 2}"
+14.015	32	14.015649	addition of acetic acid and loss of CO2. Reaction: (+C2H2O2) and (-CO2)	"{'C': 1, 'H': 2}"
+68.991	32	68.99467647	M+HCOONa[1+]	"{'C':1,'O':2, 'Na':2, 'H':1}"
+46.005	31	46.005305	addition of Phosphate and dechlorination. Reaction: (+HPO3) and (-Cl+H)	"{'H': 2, 'O': 3, 'P': 1, 'Cl': -1}"
+26.015	29	26.01565	acetylation and loss of oxygen. Reaction: (+C2H2O) and (-O)	"{'C': 2, 'H': 2}"
+61.97	29	61.975755	addition of Phosphate and defluorination. Reaction: (+HPO3) and (-F+H)	"{'H': 2, 'O': 3, 'P': 1, 'F': -1}"
+129.957	29	129.958482	addition of di-phosphate and denitrification. Reaction: (+H2P2O6) and (NO2 -> NH2)	"{'H': 4, 'O': 4, 'P': 2}"
+60.021	29	-60.0211	desmolysis	"{'C': -2, 'H': -4, 'O': -2}"
+74.0365	28	74.03678	propionylation	"{'C': 3, 'H': 6, 'O': 2}"
+23.9995	27	24	acetylation and loss of water. Reaction: (+C2H2O) and (-H2O)	{'C': 2}
+10.029	27	10.028802	addition of CO2 and dechlorination. Reaction: (+CO2) and (-Cl+H)	"{'C': 1, 'H': 1, 'O': 2, 'Cl': -1}"
+44.026	26	44.0262	hydroxyethylation	"{'C': 2, 'H': 4, 'O': 1}"
+18.0105	26	18.010565	water	"{'H': 2, 'O': 1}"
+6.0165	24	6.010565	addition of tiglyl and loss of phenyl. Reaction: (+C5H6O) and (-C6H5+H)	"{'C': -1, 'H': 2, 'O': 1}"
+3.0185	23	3.019	C isotope and reduction	"{'H': 2, '(C12)': -1, '(C13)': 1}"
+15.994	23	15.9949	oxidation	{'O': 1}
+24.044	22	24.044451	addition of acetic acid and dechlorination. Reaction: (+C2H2O2) and (-Cl+H)	"{'C': 2, 'H': 3, 'O': 2, 'Cl': -1}"
+28.0305	22	28.0313	"± C2H4, natural alkane chains such as fatty acids"	"{'C': 2, 'H': 4}"
+42.01	20	42.010564	malonylation and loss of CO2. Reaction: (+C3H2O3) and (-CO2)	"{'C': 2, 'H': 2, 'O': 1}"
+12	20	12	methylation and reduction	{'C': 1}
+27.995	20	27.99492	± CO	"{'C': 1, 'O': 1}"
+57.9585	20	57.958622	NaCl	"{'Na': 1, 'Cl': 1}"
