+++ analysis/3_get_stats_rtbins/top_frequent_delta_mz_tof_pos_20241227.tsv
@@ -1,0 +1,70 @@
+delta_mz	count_estimate	mass_signature	note	dict
+1.003	697	1.00335	C isotope	{'(C12)': -1, '(C13)': 1}
+18.01	113	18.010565	water	{'H': 2, 'O': 1}
+21.9825	105	21.981943	Sodium adduct	{'H': -1, 'Na': 1}
+14.015	83	14.015649	addition of acetic acid and loss of CO2. Reaction: (+C2H2O2) and (-CO2)	{'C': 1, 'H': 2}
+2.015	80	2.014552	2H	{'H': 2}
+1.9975	76	1.9972	M(Cl37)	{'Cl': -1, '(Cl37)': 1}
+0.501	69	0.5017	double charged with C13	{'C': -1, '(C13)': 1}
+2.0055	66	2.00425	O isotope	{'(O16)': -1, '(O18)': 1}
+28.0305	59	28.0313	± C2H4, natural alkane chains such as fatty acids	{'C': 2, 'H': 4}
+22.9845	54	22.989218	 Na	  { 'Na':1 }
+46.0055	54	46.00548	± CO+H2O (carboxylic acid)	{'C': 1, 'O': 2, 'H': 2}
+17.0265	53	17.0265	addition of ammonia. Reaction: (+NH3)	{'N': 1, 'H': 3}
+11.9995	45	12.0	methylation and reduction	{'C': 1}
+44.0255	41	44.0262	hydroxyethylation	{'C': 2, 'H': 4, 'O': 1}
+26.015	41	26.01565	acetylation and loss of oxygen. Reaction: (+C2H2O) and (-O)	{'C': 2, 'H': 2}
+15.995	40	15.99492	± O, e.g. oxidation/reduction	{'O': 1}
+16.031	39	16.0313	Methylation + reduction	{'C': 1, 'H': 4}
+3.0185	38	3.019	C isotope and reduction	{'H': 2, '(C12)': -1, '(C13)': 1}
+32.026	37	32.026215	MeOH	{'C': 1, 'H': 4, 'O': 1}
+39.993	37	39.9925	extra OH sodium adduct	{'H': 1, 'O': 1}
+27.9945	37	27.9949	addition of CO. Reaction: (+CO)	{'C': 1, 'O': 1}
+23.999	36	24.0	acetylation and loss of water. Reaction: (+C2H2O) and (-H2O)	{'C': 2}
+15.973	35	15.9738	lithium to sodium adduct	{}
+19.014	35	19.01787647	M+H2O+H[1+]	 { 'H': 3, 'O':1}"
+42.0465	35	42.04695	± C3H6, propylation	{'C': 3, 'H': 6}
+9.984	32	9.98435	addition of CO and loss of water. Reaction: (+CO) and (-H2O)	{'C': 1, 'H': -2}
+30.0105	31	30.010564	addition of acetic acid and loss of CO. Reaction: (+C2H2O2) and (-CO)	{'C': 1, 'H': 2, 'O': 1}
+56.0625	30	56.0626	± C4H8, butylation	{'C': 4, 'H': 8}
+13.979	30	13.979264	nitrification and loss of oxygen. Reaction: (NH2 -> NO2) and (-O)	{'H': -2, 'O': 1}
+60.021	29	-60.0211	desmolysis	{'C': -2, 'H': -4, 'O': -2}
+42.01	29	42.010564	malonylation and loss of CO2. Reaction: (+C3H2O3) and (-CO2)	{'C': 2, 'H': 2, 'O': 1}
+3.9945	29	3.994915	oxidative cleavage	{'O': 1, 'C': -1}
+4.9555	29	4.9554	Na+<-> NH4+, salt adduct	{'Na': 1, 'N': -1, 'H': -4}
+67.9875	28	67.987424	NaCOOH	{'C': 1, 'O': 2, 'Na': 1, 'H': 1}
+74.0185	28	74.018791	polysiloxane repeat unit	{'C': 2, 'H': 6, 'O': 1, 'Si': 1}
+83.9525	27
+17.0065	27
+61.97	27
+21.989	26
+88.052	26	88.052429	butanoic acid	{'C': 4, 'H': 8, 'O': 2}
+40.0305	26	40.0313	"(C3H6O - H2O), acetone condensation after dehydration"	{'H': 4, 'C': 3}
+7.989	25
+17.9725	25	17.974179	malonylation and loss of crotonyl. Reaction: (+C3H2O3) and (-C4H4O)	{'C': -1, 'H': -2, 'O': 2}
+41.026	25	41.026549	Acetonitrile	{'C': 2, 'H': 3, 'N': 1}
+0.9845	25	0.98402	OH <-> NH2, e.g. de-amidiation, CHNO compounds	{'O': 1, 'H': -1, 'N': -1}
+30.0465	25	30.04695	addition of C2H4 and hydrogenation. Reaction: (+C2H4) and (+H2)	{'C': 2, 'H': 6}
+72.0205	25	72.021129	lactic acid	{'C': 3, 'H': 4, 'O': 2}
+37.9555	25	37.955768	addition of Phosphate and Deacetylation. Reaction: (+HPO3) and (-C2H2O)	{'C': -2, 'H': -1, 'O': 2, 'P': 1}
+58.0415	24	58.041865	PPG repeat unit	{'C': 3, 'H': 6, 'O': 1}
+1.9795	24	1.979265	demethylation + hydroxylation (on same or differing sites)	{'O': 1, 'C': -1, 'H': -2}
+18.0305	24
+36.0205	23	36.021129	2x water	{'H': 4, 'O': 2}
+3.0005	23	2.9997	hydrolysis	{'N': -1, 'O': 1, 'H': 1}
+55.989	22	55.989829	addition of acetic acid and dehydrogenation. Reaction: (+C2H2O2) and (-H2)	{'C': 2, 'O': 2}
+3.01	22	3.010899	addition of ammonia and demethylation. Reaction: (+NH3) and (-CH2)	{'C': -1, 'H': 1, 'N': 1}
+8.9865	22
+13.012	21
+35.037	21	35.037114	addition of water and addition of ammonia. Reaction: (+H2O) and (+NH3)	{'H': 5, 'N': 1, 'O': 1}
+56.025	21	56.024872		{'C': 1, 'H': 2, 'N': 3}
+17.0305	21
+15.0185	21	15.017424	addition of glucuronic acid and loss of N-acetyl-S-cysteine. Reaction: (+C6H8O6) and (-C5H7NO3S)	{'C': 1, 'H': 1, 'N': -1, 'O': 3, 'S': -1}
+72.0575	21	72.057515	butyrylation and hydrogenation. Reaction: (+C4H6O) and (+H2)	{'C': 4, 'H': 8, 'O': 1}
+20.979	20
+47.008	20
+54.046	20	54.04695	butyrylation and loss of oxygen. Reaction: (+C4H6O) and (-O)	{'C': 4, 'H': 6}
+58.004	20	58.004137		{'N': 3, 'O': 1}
+1.019	20
+21.9735	20
+29.034	20	29.033074	addition of D-ribose and loss of cysteine. Reaction: (+C5H8O4) and (-C3H5NOS)	{'C': 2, 'H': 3, 'N': -1, 'O': 3, 'S': -1}
