+++ analysis/4_plot_rt_distri/selected_16_tof_datasets.txt
@@ -1,0 +1,16 @@
+ST001828_RPneg_NEG_ppm25_1031133743
+ST000046_HILICpos__ppm25_1031113822
+ST000726_RPneg__ppm25_1031133359
+ST000046_RPneg__ppm25_1031114022
+ST002711_HILICneg_467535 Lipkin negHILIC 6550_B4_ppm25_1031115431
+ST002711_HILICpos_467535 Lipkin posHILIC 6530b_B7_ppm25_1031114314
+ST000726_RPpos__ppm25_103113299
+MTBLS718_RPpos_LPOS_B3_ppm25_1031121232
+ST002700_RPneg_466506 Lipkin negCSH 6550_B8_ppm25_103112379
+ST001828_RPpos_POS_ppm25_1031131529
+ST001217_RPpos__ppm25_1031115140
+MTBLS718_HILICpos_HPOS_B1_ppm25_1031115857
+ST001217_HILICpos__ppm25_103111503
+MTBLS1133_RPpos_B2_ppm25_103113407
+ST002826_HILICpos__ppm25_1031121621
+MTBLS718_RPneg_LNEG_B2_ppm25_1031122629
