+++ analysis/5_comprehensive_khipu_search/elution_parameters_16studies_tof.tsv
@@ -1,0 +1,17 @@
+ID	number_good_khipus	mean_m0	stdev_m0	mean_shift	stdev_shift
+ST001828_RPneg_NEG_ppm25_1031133743	2857	7.372	1.244	0.006	0.65
+ST000046_HILICpos__ppm25_1031113822	297	9.375	2.431	-0.048	0.654
+ST000726_RPneg__ppm25_1031133359	1394	10.455	2.032	-0.023	0.721
+ST000046_RPneg__ppm25_1031114022	135	9.521	4.489	-0.147	0.6
+ST002711_HILICneg_467535 Lipkin negHILIC 6550_B4_ppm25_1031115431	665	10.597	10.521	-0.09	0.807
+ST002711_HILICpos_467535 Lipkin posHILIC 6530b_B7_ppm25_1031114314	341	10.758	2.721	-0.031	0.787
+ST000726_RPpos__ppm25_103113299	1247	10.326	2.049	-0.02	0.771
+MTBLS718_RPpos_LPOS_B3_ppm25_1031121232	1174	5.132	1.041	-0.061	0.446
+ST002700_RPneg_466506 Lipkin negCSH 6550_B8_ppm25_103112379	1183	9.348	2.867	-0.008	0.698
+ST001828_RPpos_POS_ppm25_1031131529	3621	7.389	1.069	-0.003	0.715
+ST001217_RPpos__ppm25_1031115140	1483	9.502	2.368	-0.022	0.822
+MTBLS718_HILICpos_HPOS_B1_ppm25_1031115857	1090	4.219	2.36	-0.006	0.692
+ST001217_HILICpos__ppm25_103111503	343	10.89	1.854	0.034	0.903
+MTBLS1133_RPpos_B2_ppm25_103113407	873	18.461	5.114	-0.014	0.982
+ST002826_HILICpos__ppm25_1031121621	199	12.009	2.463	-0.252	0.738
+MTBLS718_RPneg_LNEG_B2_ppm25_1031122629	437	5.069	0.795	-0.013	0.51
