+++ analysis/5_comprehensive_khipu_search/elution_parameters_45studies_orbi.tsv
@@ -1,0 +1,46 @@
+ID	number_good_khipus	mean_m0	stdev_m0	mean_shift	stdev_shift
+ST001237_HILICpos_B2_ppm5_3524314	2893	6.475	3.651	-0.033	0.613
+ST002937_HILICpos_ST002937_pos_hilic_batch_1_zip_B2_ppm5_351419	4771	8.335	10.071	-0.015	0.786
+ST002112_HILICpos_B2_ppm5_3545123	3152	8.207	7.133	-0.029	0.789
+MTBLS4187_HILICpos_HILIC_ppm5_3521045	2242	6.865	10.593	-0.034	0.604
+MTBLS3852_HILICpos__ppm5_3533848	316	4.97	19.311	-0.009	0.677
+MTBLS1465_HILICpos__ppm5_3505731	2189	10.138	16.226	-0.009	0.658
+ST001335_HILICpos__ppm5_35132731	1563	6.24	3.422	-0.063	0.633
+ST001736_RPpos__ppm5_34232333	943	8.495	5.84	-0.019	0.717
+MTBLS136_RPpos_B17_ppm5_3583756	1353	4.426	1.165	-0.007	0.535
+ST002200_RPpos_17min_B3_ppm5_3422144	3361	18.187	6.088	-0.11	0.904
+ST001181_RPpos_B11_ppm5_36163417	261	9.983	18.02	-0.067	0.896
+ST002443_RPpos_B3_ppm5_81311338	395	28.966	20.845	-0.076	1.035
+ST002118_RPpos_B1_ppm5_36115624	1257	31.071	20.266	-0.16	0.994
+ST003315_RPpos__ppm5_813163714	909	5.879	4.957	-0.004	0.481
+ST002112_RPpos_B3_ppm5_356194	5963	7.451	2.796	-0.027	0.711
+MTBLS4187_RPpos__ppm5_3522135	6053	6.292	1.414	-0.023	0.621
+ST002049_RPpos_B1_ppm5_352018	1730	30.632	9.488	-0.125	1.02
+MTBLS2245_RPpos_Method 2_B1_ppm5_81318250	1093	4.707	0.902	-0.006	0.516
+ST001430_RPpos_B1_ppm5_3423140	3348	7.86	10.38	-0.038	0.58
+MTBLS205_RPpos_B1_ppm5_3531210	1036	4.485	1.013	-0.036	0.516
+ST001243_RPpos__ppm5_35185010	1078	8.76	2.714	-0.039	0.711
+ST002521_RPpos_B3_ppm5_3662420	7204	8.894	2.17	-0.028	0.683
+MTBLS2245_HILICneg_Method 4_B5_ppm5_81318755	896	4.978	0.734	-0.022	0.519
+ST002937_HILICneg_ST002937_neg_hilic_batch_1_zip_B3_ppm5_353747	3660	8.186	9.816	-0.024	0.812
+ST002112_HILICneg_B3_ppm5_3552830	2825	7.786	7.23	0.018	0.789
+MTBLS1465_HILICneg__ppm5_35195	676	9.762	6.08	-0.032	0.735
+MTBLS205_HILICneg__ppm5_353536	719	4.445	2.347	-0.032	0.55
+ST002331_RPneg_B5_ppm5_3655219	698	7.297	7.269	-0.022	0.656
+MTBLS476_RPneg__ppm5_3551151	6783	7.792	2.1	-0.024	0.695
+MTBLS4294_RPneg__ppm5_3542818	2811	10.649	5.969	-0.041	0.777
+ST002200_RPneg_5min_B4_ppm5_3423035	802	10.238	3.211	-0.045	0.688
+ST001430_RPneg_B4_ppm5_350239	2347	7.763	4.168	-0.068	0.6
+ST002238_RPneg__ppm5_36104557	2599	10.757	4.033	-0.024	0.822
+ST002521_RPneg_B2_ppm5_3673943	4043	9.872	6.518	-0.038	0.676
+ST002112_RPneg_B2_ppm5_3572946	5879	7.726	4.679	-0.032	0.712
+ST001736_RPneg__ppm5_34232652	913	10.409	8.857	0.004	0.688
+ST002554_RPneg__ppm5_3611102	2429	10.237	9.354	0.012	0.846
+MTBLS1465_RPneg__ppm5_3514212	1742	2.378	4.884	-0.033	0.536
+MTBLS4187_RPneg__ppm5_3524019	1969	6.298	2.423	-0.015	0.565
+MTBLS205_RPneg__ppm5_3531433	1761	3.918	1.313	-0.004	0.535
+ST001240_RPneg__ppm5_35193041	1272	7.914	3.274	0.037	0.731
+ST003315_RPneg__ppm5_813164016	1155	5.992	4.789	-0.012	0.576
+ST000919_RPneg_B1_ppm5_363371	6847	5.491	1.547	-0.017	0.615
+ST001243_RPneg__ppm5_3518535	531	9.715	2.999	0.0	0.732
+ST001335_RPneg__ppm5_35133334	1072	5.932	1.246	0.01	0.521
