--- analysis/selected_45_orbi_datasets.txt
@@ -1,45 +1,0 @@
-ST001237_HILICpos_B2_ppm5_3524314
-ST002937_HILICpos_ST002937_pos_hilic_batch_1_zip_B2_ppm5_351419
-ST002112_HILICpos_B2_ppm5_3545123
-MTBLS4187_HILICpos_HILIC_ppm5_3521045
-MTBLS3852_HILICpos__ppm5_3533848
-MTBLS1465_HILICpos__ppm5_3505731
-ST001335_HILICpos__ppm5_35132731
-ST001736_RPpos__ppm5_34232333
-MTBLS136_RPpos_B17_ppm5_3583756
-ST002200_RPpos_17min_B3_ppm5_3422144
-ST001181_RPpos_B11_ppm5_36163417
-ST002443_RPpos_B3_ppm5_81311338
-ST002118_RPpos_B1_ppm5_36115624
-ST003315_RPpos__ppm5_813163714
-ST002112_RPpos_B3_ppm5_356194
-MTBLS4187_RPpos__ppm5_3522135
-ST002049_RPpos_B1_ppm5_352018
-MTBLS2245_RPpos_Method 2_B1_ppm5_81318250
-ST001430_RPpos_B1_ppm5_3423140
-MTBLS205_RPpos_B1_ppm5_3531210
-ST001243_RPpos__ppm5_35185010
-ST002521_RPpos_B3_ppm5_3662420
-MTBLS2245_HILICneg_Method 4_B5_ppm5_81318755
-ST002937_HILICneg_ST002937_neg_hilic_batch_1_zip_B3_ppm5_353747
-ST002112_HILICneg_B3_ppm5_3552830
-MTBLS1465_HILICneg__ppm5_35195
-MTBLS205_HILICneg__ppm5_353536
-ST002331_RPneg_B5_ppm5_3655219
-MTBLS476_RPneg__ppm5_3551151
-MTBLS4294_RPneg__ppm5_3542818
-ST002200_RPneg_5min_B4_ppm5_3423035
-ST001430_RPneg_B4_ppm5_350239
-ST002238_RPneg__ppm5_36104557
-ST002521_RPneg_B2_ppm5_3673943
-ST002112_RPneg_B2_ppm5_3572946
-ST001736_RPneg__ppm5_34232652
-ST002554_RPneg__ppm5_3611102
-MTBLS1465_RPneg__ppm5_3514212
-MTBLS4187_RPneg__ppm5_3524019
-MTBLS205_RPneg__ppm5_3531433
-ST001240_RPneg__ppm5_35193041
-ST003315_RPneg__ppm5_813164016
-ST000919_RPneg_B1_ppm5_363371
-ST001243_RPneg__ppm5_3518535
-ST001335_RPneg__ppm5_35133334
