--- analysis/top_frequent_delta_mz_orbi_neg-20241116.tsv
@@ -1,109 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0033	3742	1.0033	['1.0034', 'M(C13)', "{'C': -1, '(C13)': 1}"]
-67.9874	762	67.9873	['67.987424', 'NaCOOH', "{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"]
-14.0156	614	14.0155	['14.01565', "± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)", "{'C': 1, 'H': 2}"]
-2.0156	510	2.0156	['2.01565', '± 2H, opening or forming of double bond', "{'H': 2}"]
-82.0029	401	82.0029	['82.0030 C2HF3']
-1.9969	376	1.9969	['1.9970', 'M(Cl37)', "{'Cl': -1, '(Cl37)': 1}"]
-15.9948	371	15.9948	['15.99492', '± O, e.g. oxidation/reduction', "{'O': 1}"]
-43.9898	356	43.9897	['43.98983', '± CO2', "{'C': 1, 'O': 2}"]
-18.0105	332	18.0104	['-18.010565', 'H2O', "{'H': -2, 'O': -1}"]
-11.9999	324	11.9999	['11.99827646677', 'M+H+Na[2+]', "{'H': 1, 'Na': 1}"]
-30.0105	316
-26.0156	313	26.0155	[' C2H2']
-46.0054	308	46.0053	['-46.005479', 'H2O+CO', "{'C': -1, 'H': -2, 'O': -2}"]
-28.0312	299	28.0312	['28.0313', '± C2H4, natural alkane chains such as fatty acids', "{'C': 2, 'H': 4}"]
-44.0261	280
-21.9819	270	21.9818	['21.98194', 'Na+<-> H+, salt adduct', "{'Na': 1, 'H': -1}"]
-27.9948	267	27.9948	['27.99492', '± CO', "{'C': 1, 'O': 1}"]
-1.9957	255	1.9957	['1.99566', 'F <-> OH, halogen exchange with hydroxy group (typically -F + OH)', "{'F': 1, 'O': -1, 'H': -1}"]
-10.0287	254
-66.984	247	66.984	['-66.98012353323', 'M-HCOONa+H[1+]', "{'C': -1, 'O': -2, 'Na': -1}"]
-23.9999	238	23.9999	['23.996494', 'H+Na', "{'Na': 1, 'H': 1}"]
-2.0061	231	2.0062	['2.0067', '2*(C13)']
-13.9792	230	13.9791	['13.97927', 'O <-> 2H, e.g. Oxidation follwed by H2O elimination', "{'H': -2, 'O': 1}"]
-42.0105	229	42.0104	['42.01057', '± COCH2', "{'C': 2, 'O': 1, 'H': 2}"]
-135.9749	217	135.9748	['135.974848', '2X NaCOOH', "{'C': 2, 'O': 4, 'H': 2, 'Na': 2}"]
-16.0312	214
-60.021	209
-1.9978	201	1.9978	['1.9972', 'M(Cl37)', "{'Cl': -1, '(Cl37)': 1}"]
-2.0042	200	2.0062	['2.0067', '2*(C13)']
-1.9792	196	1.9792	['1.97975', 'K+ <-> Cl-+2H2+, salt adduct', "{'K': -1, 'Cl': 1, 'H': 2}"]
-15.9738	192	15.9738	['15.97394', 'Na+<-> K+, salt adduct', "{'Na': -1, 'K': 1}"]
-3.9948	190	3.9948	['(Ca44)']
-32.0261	181	32.0261	['32.026215', 'MeOH', "{'C': 1, 'H': 4, 'O': 1}"]
-57.9586	176	57.9585	['57.958622', 'NaCl', "{'Na': 1, 'Cl': 1}"]
-4.0312	176
-9.9843	173
-46.0418	166
-15.0189	163
-21.9842	161	21.9818	['21.98194', 'Na+<-> H+, salt adduct', "{'Na': 1, 'H': -1}"]
-58.0418	159	58.0417	['58.04187', '"+C3H6O, acetone condensation"', "{'C': 3, 'H': 6, 'O': 1}"]
-3.0189	155	3.0189	['3.021828', '3H', "{'H': 3}"]
-68.9908	155	68.9907	['68.99467646676999', 'M+HCOONa[1+]', "{'C': 1, 'O': 2, 'Na': 2, 'H': 1}"]
-42.0469	154	42.0468	['42.04695', '± C3H6, propylation', "{'C': 3, 'H': 6}"]
-56.0261	152
-20.0261	148
-13.0122	148
-58.0054	148	58.0053	['58.00548', '± CO2CH2', "{'C': 2, 'O': 2, 'H': 2}"]
-27.0189	147
-31.9898	146	31.9897	['31.98983', '± 2O, oxygen loss', "{'O': 2}"]
-70.0418	143
-30.0469	142
-164.0058	142
-39.9949	141	39.9948	['39.99248', 'Na+ <-> -H2O+H', "{'Na': 1, 'O': 1, 'H': 1}"]
-25.9792	141
-6.0169	141
-47.0088	140
-1.0122	139	1.0077	['-1.00727646677', '-H[-]', "{'H': -1}"]
-12.0363	137
-72.0574	135
-80.9995	135
-38.0156	135
-40.0312	134	40.0312	['40.0313', '"(C3H6O - H2O), acetone condensation after dehydration"', "{'H': 4, 'C': 3}"]
-54.0105	132
-10.0206	132
-8.005	127
-6.0105	126
-50.0156	125
-74.0367	124
-72.0211	122
-62.0003	121
-8.0317	120
-34.0054	117
-11.9635	115
-29.9741	115	29.9741	['29.97418', 'NO2 <-> NH2, nitro compounds', "{'H': -2, 'O': 2}"]
-35.9766	115	35.9766	['35.9767', 'Cl-+2H2+ <-> H+, salt adduct', "{'Cl': 1, 'H': 1}"]
-88.0524	115
-56.0625	114	56.0624	['56.0626', '± C4H8, butylation', "{'C': 4, 'H': 8}"]
-2.9996	113	2.9996	['3.00307646677', 'M(S34)+H[1+]', "{'S': -1, '(S34)': 1, 'H': 1}"]
-52.0312	112
-48.021	111
-78.0164	111
-0.9935	111	0.9935	['0.98992353323', 'M(Cl37)-H[-]', "{'Cl': -1, '(Cl37)': 1, 'H': -1}"]
-18.0469	111
-68.0261	110
-35.9999	110
-16.9901	110
-29.9895	109
-97.9769	109	97.9768	['97.9769', '± H3PO4, phosphorous compounds', "{'O': 4, 'H': 3, 'P': 1}"]
-62.0367	108
-54.0469	108
-1.9946	107	1.9945	['1.99566', 'F <-> OH, halogen exchange with hydroxy group (typically -F + OH)', "{'F': 1, 'O': -1, 'H': -1}"]
-34.0418	107
-0.9951	106	0.9935	['0.98992353323', 'M(Cl37)-H[-]', "{'Cl': -1, '(Cl37)': 1, 'H': -1}"]
-67.9897	106	67.9896	['67.987424', 'NaCOOH', "{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"]
-60.0574	106
-61.9703	106
-44.9931	105
-0.5016	105	0.5016	['0.5017', 'double charged with C13', "{'C': -1, '(C13)': 1}"]
-52.0134	104
-5.9894	104
-82.0053	103	82.0029	['82.0030 C2HF3']
-24.0443	103
-77.9772	102
-86.0367	102
-90.0315	101
-7.9686	101
-41.0265	101	41.0264	['41.026549', 'Acetonitrile', "{'C': 2, 'H': 3, 'N': 1}"]
-83.0063	101
