--- analysis/top_frequent_delta_mz_orbi_pos-20241116.tsv
@@ -1,144 +1,0 @@
-delta_mz	count_estimate	mass_signature	note	dict
-1.0033	3656	1.0033	['1.0034', 'M(C13)', "{'C': -1, '(C13)': 1}"]
-14.0155	792	14.0155	['14.01565', "± CH2, alkane chains, waxes, fatty acids, methylation; or '-[C3H6ON] <-> -[C2H4ON], acrylamide versus iodoacetamide in cysteine alkylation (gels)", "{'C': 1, 'H': 2}"]
-18.0104	780	18.0104	['-18.010565', 'H2O', "{'H': -2, 'O': -1}"]
-2.0156	634	2.0156	['2.01565', '± 2H, opening or forming of double bond', "{'H': 2}"]
-21.9818	590	21.9818	['21.98194', 'Na+<-> H+, salt adduct', "{'Na': 1, 'H': -1}"]
-44.0261	585
-28.0312	550	28.0312	['28.0313', '± C2H4, natural alkane chains such as fatty acids', "{'C': 2, 'H': 4}"]
-15.9948	420	15.9948	['15.99492', '± O, e.g. oxidation/reduction', "{'O': 1}"]
-17.0264	404	17.0264	['-17.026549', 'NH3', "{'N': -1, 'H': -3}"]
-26.0155	392	26.0155	[' C2H2']
-27.9948	385	27.9948	['27.99492', '± CO', "{'C': 1, 'O': 1}"]
-0.5016	383	0.5016	['0.5017', 'double charged with C13', "{'C': -1, '(C13)': 1}"]
-11.9999	380	11.9999	['11.99827646677', 'M+H+Na[2+]', "{'H': 1, 'Na': 1}"]
-30.0104	328
-32.026	311	32.0261	['32.026215', 'MeOH', "{'C': 1, 'H': 4, 'O': 1}"]
-42.0104	301	42.0104	['42.01057', '± COCH2', "{'C': 2, 'O': 1, 'H': 2}"]
-67.9872	295	67.9873	['67.987424', 'NaCOOH', "{'C': 1, 'O': 2, 'Na': 1, 'H': 1}"]
-13.9791	287	13.9791	['13.97927', 'O <-> 2H, e.g. Oxidation follwed by H2O elimination', "{'H': -2, 'O': 1}"]
-23.9999	281	23.9999	['23.996494', 'H+Na', "{'Na': 1, 'H': 1}"]
-16.0312	279
-42.0468	278	42.0468	['42.04695', '± C3H6, propylation', "{'C': 3, 'H': 6}"]
-46.0053	277	46.0053	['-46.005479', 'H2O+CO', "{'C': -1, 'H': -2, 'O': -2}"]
-88.0522	277
-41.0264	262	41.0264	['41.026549', 'Acetonitrile', "{'C': 2, 'H': 3, 'N': 1}"]
-22.9852	257	22.9852	['22.989218', 'Na', "{'Na': 1}"]
-15.9738	255	15.9738	['15.97394', 'Na+<-> K+, salt adduct', "{'Na': -1, 'K': 1}"]
-30.0468	238
-4.0312	231
-60.0209	230
-56.0624	229	56.0624	['56.0626', '± C4H8, butylation', "{'C': 4, 'H': 8}"]
-9.9842	226
-3.9948	218	3.9948	['(Ca44)']
-58.0417	217	58.0417	['58.04187', '"+C3H6O, acetone condensation"', "{'C': 3, 'H': 6, 'O': 1}"]
-1.9791	217	1.9792	['1.97975', 'K+ <-> Cl-+2H2+, salt adduct', "{'K': -1, 'Cl': 1, 'H': 2}"]
-4.9553	214	4.9553	['4.9554', 'Na+<-> NH4+, salt adduct', "{'Na': 1, 'N': -1, 'H': -4}"]
-2.0065	206	2.0062	['2.0067', '2*(C13)']
-19.0138	205	19.0138	['-19.01787646677', 'M-H2O-H[-]', "{'H': -3, 'O': -1}"]
-36.021	201
-46.0417	200
-15.0189	200
-40.0311	188	40.0312	['40.0313', '"(C3H6O - H2O), acetone condensation after dehydration"', "{'H': 4, 'C': 3}"]
-1.9969	188	1.9969	['1.9970', 'M(Cl37)', "{'Cl': -1, '(Cl37)': 1}"]
-13.0122	187
-17.0071	187	17.0071	['-17.00332353323', 'M-H2O+H[1+]', "{'H': -1, 'O': -1}"]
-20.0261	183
-3.0189	181	3.0189	['3.021828', '3H', "{'H': 3}"]
-45.0577	175
-72.0573	172
-58.0052	171	58.0053	['58.00548', '± CO2CH2', "{'C': 2, 'O': 2, 'H': 2}"]
-56.026	171
-21.9842	170	21.9818	['21.98194', 'Na+<-> H+, salt adduct', "{'Na': 1, 'H': -1}"]
-10.0206	164
-31.9897	164	31.9897	['31.98983', '± 2O, oxygen loss', "{'O': 2}"]
-54.0468	163
-43.0227	161
-12.0363	159
-43.9897	159	43.9897	['43.98983', '± CO2', "{'C': 1, 'O': 2}"]
-74.0365	154
-72.0209	153
-1.0122	153	1.0077	['-1.00727646677', '-H[-]', "{'H': -1}"]
-60.0573	152
-1.0008	152	0.9994	['1.0034', 'M(C13)', "{'C': -1, '(C13)': 1}"]
-22.013	150
-15.0108	150
-39.9924	149	39.9924	['39.99248', 'Na+ <-> -H2O+H', "{'Na': 1, 'O': 1, 'H': 1}"]
-2.0042	149	2.0062	['2.0067', '2*(C13)']
-70.0781	148
-70.0416	147
-39.9947	144	39.9948	['39.99248', 'Na+ <-> -H2O+H', "{'Na': 1, 'O': 1, 'H': 1}"]
-45.0294	144
-74.0729	143
-1.998	141	1.9978	['1.9972', 'M(Cl37)', "{'Cl': -1, '(Cl37)': 1}"]
-8.005	141
-0.9839	140	0.9839	['0.98402', 'OH <-> NH2, e.g. de-amidiation, CHNO compounds', "{'O': 1, 'H': -1, 'N': -1}"]
-34.0053	140
-38.0155	140
-0.9969	139	0.9994	['1.0034', 'M(C13)', "{'C': -1, '(C13)': 1}"]
-25.9791	139
-35.037	136	35.037	['-35.037114', 'NH3+H2O', "{'H': -5, 'O': -1, 'N': -1}"]
-27.0108	136	27.0108	['27.0109', '± HCN, nitrile compounds', "{'N': 1, 'H': 1, 'C': 1}"]
-27.0278	135
-29.0345	135
-6.0104	134
-18.0468	134
-48.0209	133
-132.0785	133
-52.0311	132
-44.0624	128
-37.9557	127	37.9557	['37.955882', 'K-H', "{'K': 1, 'H': -1}"]
-0.9951	126	0.9935	['0.98992353323', 'M(Cl37)-H[-]', "{'Cl': -1, '(Cl37)': 1, 'H': -1}"]
-62.0366	125
-54.0104	125
-14.0519	124
-12.9951	124
-102.0679	124	102.0679	['C5H10O2']
-29.974	123	29.9741	['29.97418', 'NO2 <-> NH2, nitro compounds', "{'H': -2, 'O': 2}"]
-82.0029	122	82.0029	['82.0030 C2HF3']
-84.0937	122
-34.0417	118
-50.0155	117
-11.9635	116
-3.0108	116
-57.9585	116	57.9585	['57.958622', 'NaCl', "{'Na': 1, 'Cl': 1}"]
-1.9957	116	1.9957	['1.99566', 'F <-> OH, halogen exchange with hydroxy group (typically -F + OH)', "{'F': 1, 'O': -1, 'H': -1}"]
-23.0158	115
-86.0365	115
-29.0264	114
-59.0369	113	59.0369	['59.03711', '± C2H5NO, N-acetyl or cut from side chain of Asn in peptides', "{'C': 2, 'H': 5, 'N': 1, 'O': 1}"]
-50.0366	113
-17.9741	112
-43.042	112
-7.9686	111
-42.0297	111
-1.0077	111	1.0077	['-1.00727646677', '-H[-]', "{'H': -1}"]
-6.0468	110
-86.0729	110
-35.9998	110
-16.0231	110	16.0231	['-16.019223533229997', 'M-NH3+H[1+]', "{'N': -1, 'H': -2}"]
-58.078	110
-26.0519	109
-27.0189	109
-24.0363	109
-18.0298	108	18.0298	['18.033823', 'NH4+', "{'N': 1, 'H': 4}"]
-32.0625	108
-135.9745	106	135.9748	['135.974848', '2X NaCOOH', "{'C': 2, 'O': 4, 'H': 2, 'Na': 2}"]
-5.9893	106
-13.0315	105
-68.0624	105
-14.9914	105	14.9914	['14.987633533230001', 'M-H+O[-]', "{'H': -1, 'O': 1}"]
-64.0522	105
-2.9996	104	2.9996	['3.00307646677', 'M(S34)+H[1+]', "{'S': -1, '(S34)': 1, 'H': 1}"]
-68.026	103
-22.0417	103
-61.0526	103
-8.0261	103
-76.0522	102
-55.9896	102
-10.0287	102
-31.042	102
-48.0573	101
-25.0122	101
-45.0213	101
-19.9685	100
